db and seed cleanup and reset

+++ db/migrate/20200804202635_create_regions.rb
@@ -2,8 +2,8 @@ class CreateRegions < ActiveRecord::Migration[6.0]
   def change
     create_table :regions do |t|
       t.string :name
-      t.integer :lat, precision: 10, scale: 6
-      t.integer :lng, precision: 10, scale: 6
+      t.decimal :lat, precision: 10, scale: 6
+      t.decimal :lng, precision: 10, scale: 6
       t.integer :zoom
 
       t.timestamps
